Second attempt to fix mouse events.
Do not apply the bottom border to `hoverDiv`. The bottom border was calculated once only, and is likely to be incorrect if the browser window is resized. However, this change means that the `hoverDiv` is larger than necessary.

To counter this problem `getPositionFromMouseCoords` now has the return type `Position|undefined` and returns undefined if the mouse position is outside the fretboard by more than half the string separation.

The calculation in `getPositionFromMouseCoords` has been simplified. It is now relative only to the stringsGroup and does not rely on the topPadding or LeftPadding.